fix: изменить тип поля confirmedRequests

+++ ewm-service/src/main/java/ru/practicum/explorewithme/event/model/EventEntity.java
@@ -56,7 +56,7 @@ public class EventEntity {
   private EventStatus state;
   private LocalDateTime createdOn;
   private LocalDateTime publishedOn;
-  private Long confirmedRequests;
+  private Integer confirmedRequests;
 
   public EventEntity(Long id) {
     this.id = id;
